LayoutTests/http/tests/app-privacy-report/user-attribution-post-request.html

<!DOCTYPE html><!-- webkit-test-runner [ isAppInitiated=false ] -->
<html>
<head>
<meta charset="utf-8">
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<p>Tests non app initiated requests get marked for POST requests.</p>
<form id="testForm" action="resources/user-initiated-post.py" target="_blank" rel="noreferrer" method="post">
<input type="submit" value="Test"></input>
<form>
<script>
if (window.testRunner)
    testRunner.waitUntilDone();

document.forms[0].submit();

</script>
</body>
</html>
